DRY AutoScopeOnPrepare
