+++ sqlitecluster/SQLiteCore.cpp
@@ -1,3 +1,4 @@
+#include <libstuff/AutoScopeOnPrepare.h>
 #include <libstuff/libstuff.h>
 #include "SQLiteCore.h"
 #include "SQLite.h"
@@ -5,28 +6,6 @@
     
 SQLiteCore::SQLiteCore(SQLite& db) : _db(db)
 { }
-
-// RAII-style mechanism for automatically setting and unsetting an on prepare handler
-class AutoScopeOnPrepare {
-  public:
-    AutoScopeOnPrepare(bool enable, SQLite& db, void (*handler)(SQLite& _db, int64_t tableID)) : _enable(enable), _db(db), _handler(handler) {
-        if (_enable) {
-            _db.setOnPrepareHandler(_handler);
-            _db.enablePrepareNotifications(true);
-        }
-    }
-    ~AutoScopeOnPrepare() {
-        if (_enable) {
-            _db.setOnPrepareHandler(nullptr);
-            _db.enablePrepareNotifications(false);
-        }
-    }
-
-  private:
-    bool _enable;
-    SQLite& _db;
-    void (*_handler)(SQLite& _db, int64_t tableID);
-};
 
 bool SQLiteCore::commit(const string& description, uint64_t& commitID, string& transactionHash,
                         bool needsPluginNotification, void (*notificationHandler)(SQLite& _db, int64_t tableID)) {
